finish adding user endpoints and snowflake for id

+++ internal/community/models.go
@@ -3,7 +3,7 @@ package community
 import "time"
 
 type Community struct {
-	ID           int       `db:"community_id" json:"id"`
+	ID           int64       `db:"community_id" json:"id"`
 	Name         string    `db:"name" json:"name"`
 	Introduction string    `db:"introduction" json:"introduction"`
 	Create_time  time.Time `db:"create_time" json:"create_time"`
@@ -11,7 +11,7 @@ type Community struct {
 }
 
 type NewCommunity struct {
-	ID           int    `json:"id" validate:"required"`
+	ID           int64    `json:"id" validate:"required"`
 	Name         string `json:"name" validate:"required"`
 	Introduction string `json:"introduction" validate:"required"`
 }
