+++ internal/post/models.go
@@ -3,7 +3,7 @@ package post
 import "time"
 
 type Post struct {
-	ID          int       `db:"post_id" json:"id" `
+	ID          int64       `db:"post_id" json:"id" `
 	Title       string    `db:"title" json:"title"`
 	Content     string    `db:"content" json:"content"`
 	Author      int       `db:"author_id" json:"author_id"`
@@ -13,7 +13,7 @@ type Post struct {
 }
 
 type NewPost struct {
-	ID        int    `db:"post_id" json:"id" validate:"required"`
+	ID        int64    `db:"post_id" json:"id" validate:"required"`
 	Title     string `db:"title" json:"title" validate:"required"`
 	Content   string `db:"content" json:"content" validate:"required"`
 	Author    int    `db:"author_id" json:"author_id" validate:"required"`
@@ -21,7 +21,7 @@ type NewPost struct {
 }
 
 type UpdatePost struct {
-	ID        int    `db:"post_id" json:"id" validate:"required"`
+	ID        int64    `db:"post_id" json:"id" validate:"required"`
 	Title     string `db:"title" json:"title" validate:"omitempty"`
 	Content   string `db:"content" json:"content" validate:"omitempty"`
 }
